requirements-dev.txt
flake8-docstrings==1.5.0
flake8-import-order==0.18.1
flake8==3.8.3
httpbin==0.7.0
mypy==0.782
pep8-naming==0.11.1
pylint==2.6.0
pytest==6.0.1
recommonmark==0.6.0
requests==2.24.0
Sphinx==3.2.1
sphinx_rtd_theme
tox==3.20.0
twine==3.2.0
wheel==0.35.1
RPi.GPIO==0.7.0